Correcion de consulta obtener perfil empresa

+++ src/main/java/com/example/proyecto_analisis/models/dto/VistaPerfilEmpresa_DTO.java
@@ -22,7 +22,7 @@ public class VistaPerfilEmpresa_DTO {
 
     private String nombreIndustria;
 
-    private Long ofertasPublicadas;
+    private int ofertasPublicadas;
     
     private String pais;
